refactor(workflows): split workflows.py into catalog + runner + facade
- workflows_catalog.py: NodeContext, NODE_TYPES registry, BUILTIN_STASHES,
  source-descriptor helpers, typed-port ADAPTERS (imports config/utils only;
  late-imports files/video inside adapters)
- workflows_runner.py: status constants, trigger types, topo_order,
  bind_participant, result sidecars, compute_resume_plan, WorkflowRunner
- workflows.py keeps the executors + collection machinery + the import-time
  wiring (mutating the shared NODE_TYPES dict imported from the catalog) and
  re-exports every public + test-touched name, so workflows.NAME and the
  monkeypatch.setitem(workflows.NODE_TYPES/ADAPTERS, ...) test seams are
  unchanged. Import DAG: catalog -> runner -> workflows; import workflows
  (not the siblings) before running graphs.
- pyproject py-modules + ARCHITECTURE.md rows + BACKEND-SPLITS-PLAN.md status

Zero workflows test edits; full suite green (2214 passed).
