+++ agents/ARCHITECTURE.md
@@ -33,7 +33,9 @@
 | [screenspace_worker.py](screenspace_worker.py) | `ScreenspaceWorker` background task queue (threading, pause/resume/cancel, multi-video, heatmaps, events). Imports tools, heatmap, manifest, multitool, frames |
 | [screenspace_preview.py](screenspace_preview.py) | Model-view preview renderer: `build_preview` (labeled multi-panel composite) + `build_overlay_layer` (single region/frame layer, catalogued in `OVERLAY_LAYERS`) show what each tool's CV pipeline actually operates on. Must mirror the real scan's preprocessing (e.g. binarized template mask via `_prepare_template`). Pure cv2/numpy leaf; imports primitives + ocr; consumed only by `screenspace_server.py`'s `/api/preview` routes |
 | [screenspace_server.py](screenspace_server.py) | Screenspace Flask REST API: region CRUD, video frame extraction, task queue management, results retrieval |
-| [workflows.py](workflows.py) | Workflows node-graph engine: the `NODE_TYPES` registry + per-node `_EXECUTORS`, typed-port `ADAPTERS`, collection-algebra control nodes, `WorkflowRunner` (topo-sorted DAG execution, per-node notes + result sidecars), and `BUILTIN_STASHES`. Orchestration-only. Reuses pipeline/screenspace/transcripts/viewer, never reimplements them |
+| [workflows.py](workflows.py) | Workflows node-graph engine: the per-node `_EXECUTORS` + collection-algebra control nodes, the import-time wiring that attaches `execute` keys to the shared `NODE_TYPES` dict, and the re-export facade over `workflows_catalog`/`workflows_runner` (import DAG: catalog → runner → workflows; `import workflows` before running graphs). Orchestration-only. Reuses pipeline/screenspace/transcripts/viewer, never reimplements them |
+| [workflows_catalog.py](workflows_catalog.py) | Declarative workflows data: `NodeContext`, the `NODE_TYPES` registry (typed ports + param schema), `BUILTIN_STASHES`, source-descriptor helpers, and the typed-port `ADAPTERS` table. Imports only config/utils (late-imports files/video inside adapters); unwired until workflows.py's import-time wiring runs |
+| [workflows_runner.py](workflows_runner.py) | Workflows run engine: run/node status constants, trigger types, `topo_order`, `bind_participant`, per-node result sidecars, `compute_resume_plan`, and `WorkflowRunner` (topo-sorted DAG execution). Reads `NODE_TYPES[...]["execute"]`/`ADAPTERS` only at call time, after the facade's wiring |
 | [workflows_server.py](workflows_server.py) | Workflows Flask blueprint: blueprint/stash CRUD, run + whole-study batch lifecycle (SSE + polling), per-node result sidecars, and the auto-run trigger daemon (single-armed per trigger type: new video / transcript complete / scan complete) |
 | [composer_server.py](composer_server.py) | Composer Flask blueprint: source-video cutting timeline. Participant/part discovery (stitched multi-part timelines), `composer_manifest.json` CRUD (cut pairs, marker trims, annotations, UI state), PIL annotation rendering, and annotated screenshot/GIF/burn exports (ffmpeg `overlay`, never `drawtext`). Cut pairs feed generation through Studio's `/api/generate-intake`; never writes the spreadsheet |
 | [overview.py](overview.py) | Overview frontend: per-participant feature-matrix builders for the Map similarity space (`build_feature_matrix` + the four `build_*_features` group builders; observations fed by sheet timestamps via the `configure()`-injected getter), the `/overview/` blueprint, and the convergence-offsets + friction-moments routes. Pure-Python (no numpy); the client does z-scoring/PCA/outlier math |
