+++ workflows.py
@@ -44,22 +44,75 @@ blueprint once per arriving participant/completion, bound via
 calls ``transcripts.transcribe_video`` directly and never writes the
 transcripts manifest, so a transcript-complete-triggered graph containing a
 Transcribe node cannot re-fire itself.
+
+This module is now the executors + wiring half and the re-export facade:
+the declarative catalog (``NodeContext``, ``NODE_TYPES``, ``BUILTIN_STASHES``,
+``ADAPTERS``) lives in ``workflows_catalog`` and the run engine
+(``WorkflowRunner``, ``topo_order``, statuses, triggers, sidecars, resume) in
+``workflows_runner``; ``import workflows; workflows.NAME`` keeps resolving
+every public name — and the private names the test suite reaches for — from
+their new homes. The import-time wiring below mutates the *shared*
+``NODE_TYPES`` dict imported from ``workflows_catalog`` (never a copy: tests
+patch node executors via ``workflows.NODE_TYPES`` and rely on the identity).
+Re-binding a name here only rebinds it on the facade — to stub a seam in a
+test, patch the owning sibling module.
 """
 
 from __future__ import annotations
 
-import copy
-import json
-import os
-import threading
-import time
-from dataclasses import dataclass, field
-from datetime import datetime, timezone
 from pathlib import Path
-from typing import Any, Callable, NotRequired, TypedDict, cast
+from typing import Any, Callable, cast
 
 import config
 import utils
+
+# Catalog names the executors + wiring below use directly (also part of the
+# ``workflows.NAME`` facade surface, like everything imported here).
+from workflows_catalog import (
+    NODE_TYPES,
+    NodeContext,
+    ParamSpec,
+    _SS_DETECTOR_SPECS,
+    _SS_REFERENCE_DETECTORS,
+    _clip_source_filename,
+    _source_descriptor,
+)
+
+# Facade re-exports — moved names that workflows_server and the test suite keep
+# reaching as ``workflows.NAME``.
+from workflows_catalog import (  # noqa: F401
+    ADAPTERS,
+    BUILTIN_STASHES,
+    NodeType,
+    Port,
+    serialize_adapters,
+    serialize_catalog,
+)
+from workflows_runner import TRIGGER_TYPE_IDS  # used by load_workflows_manifest
+from workflows_runner import (  # noqa: F401
+    NODE_STATUS_COMPLETED,
+    NODE_STATUS_FAILED,
+    NODE_STATUS_QUEUED,
+    NODE_STATUS_RUNNING,
+    NODE_STATUS_SKIPPED,
+    NOTE_NODE_TYPE,
+    RUN_STATUS_CANCELLED,
+    RUN_STATUS_COMPLETED,
+    RUN_STATUS_FAILED,
+    RUN_STATUS_QUEUED,
+    RUN_STATUS_RUNNING,
+    TRIGGER_TYPES,
+    WorkflowCycleError,
+    WorkflowRunner,
+    _inspectable_result,
+    bind_participant,
+    blueprint_participant_nodes,
+    compute_resume_plan,
+    inspectable_sidecar_view,
+    run_results_dir,
+    topo_order,
+    write_node_sidecar,
+)
 
 
 def empty_workflows_manifest() -> dict[str, list[Any]]:
@@ -133,1375 +186,6 @@ def save_workflows_manifest(
 
 
 # ---------------------------------------------------------------------------
-# Execution context
-# ---------------------------------------------------------------------------
-
-
-@dataclass
-class NodeContext:
-    """Run-wide context handed to every node executor.
-
-    Carries launch context (output dir + optional sheet) plus the uniform
-    progress and cancellation seams. Executors forward ``cancel_flag`` to
-    scan/clip callees and ``cancel_event`` to thinking-agent callees — the dual
-    contract the backend functions already expose (see
-    ``plans/archive/WORKFLOWS-PLAN.md`` and AGENTS.md). ``WorkflowRunner`` builds
-    one ``NodeContext`` per run; tests construct it directly.
-    """
-
-    input_dir: Path
-    output_dir: Path
-    sheet_context: Any = None
-    worksheet: Any = None
-    on_progress: Callable[[float], None] = lambda _fraction: None
-    cancel_event: threading.Event = field(default_factory=threading.Event)
-
-    def cancel_flag(self) -> bool:
-        """Predicate form of the cancel signal (scan/clip callees want this)."""
-        return self.cancel_event.is_set()
-
-    def resolve_videos(self, participant: str) -> list[str]:
-        """Return a participant's ordered source video path(s), or ``[]`` if none.
-
-        Multi-part participants (a session split across numbered files) return all
-        parts in timeline order — mirrors ``utils.discover_participant_videos``.
-        """
-        for entry in utils.discover_participant_videos():
-            if entry.get("id") == participant and entry.get("has_video"):
-                return list(entry["video_paths"])
-        return []
-
-
-# ---------------------------------------------------------------------------
-# Node catalog (declarative single source of truth)
-# ---------------------------------------------------------------------------
-#
-# Modelled on ``thinking_agents.AGENTS``: a data-driven, enumerable registry
-# the frontend renders generically. Each node carries the typed ports the DAG
-# needs (the wire vocabulary lives in plans/archive/WORKFLOWS-PLAN.md). Adding
-# a node is "append a NodeType + an executor" (executors section below), zero
-# frontend edits. ``serialize_catalog`` strips ``execute`` for the JSON endpoint.
-
-
-class Port(TypedDict):
-    """A typed input/output socket on a node. ``optional`` inputs may be unwired."""
-
-    name: str
-    type: str
-    optional: NotRequired[bool]
-
-
-class ParamSpec(TypedDict):
-    """A node parameter the frontend renders an editor for."""
-
-    name: str
-    type: str  # number | string | enum | bool | participant | region | ...
-    default: Any
-    label: NotRequired[str]
-    choices: NotRequired[list[Any]]
-    min: NotRequired[float]
-    max: NotRequired[float]
-    # P5: an empty value here is a guaranteed no-op/failure — the pre-run
-    # validation panel surfaces it as an error and disables Run.
-    required: NotRequired[bool]
-
-
-class NodeType(TypedDict):
-    """One node in the catalog. ``execute`` is bound in the executors section below."""
-
-    id: str
-    label: str
-    description: str  # one-line palette tooltip / on-card help text
-    domain: str  # artifact | screenspace | transcript | thinking | control
-    category: str  # human-facing palette group label
-    inputs: list[Port]
-    outputs: list[Port]
-    params: list[ParamSpec]
-    requires: list[str]  # subset of {"sheet", "videoDir"}
-    execute: NotRequired[Callable[..., dict[str, Any]]]
-    # Hidden from the palette but kept in the catalog (e.g. the per-detector
-    # ss_<tool> nodes, which the unified Detect node + Multitool read for specs).
-    hidden: NotRequired[bool]
-    # Whether a ss_<tool> detector can be a Multitool chain step (the frontend
-    # derives its step-type list from this flag — see _MULTITOOL_STEP_TOOLS).
-    multitoolStep: NotRequired[bool]
-
-
-# Shared by the three Ollama thinking nodes: a free-text override of the model
-# name (blank → the configured default). A ``string`` rather than ``enum`` because
-# the installed Ollama models are environment-specific and not known server-side.
-_OLLAMA_MODEL_PARAM: ParamSpec = {
-    "name": "model",
-    "type": "string",
-    "default": "",
-    "label": "Ollama model (blank = default)",
-}
-
-
-# Curated v1 node set (plans/archive/WORKFLOWS-PLAN.md). Keyed by id so the
-# frontend can both iterate (palette) and look up a placed node's type.
-NODE_TYPES: dict[str, NodeType] = {
-    # ---- Sources ----
-    "video_source": {
-        "id": "video_source",
-        "label": "Video Source",
-        "description": "A participant's source video, picked from the video directory.",
-        "domain": "artifact",
-        "category": "Source",
-        "inputs": [],
-        "outputs": [
-            {"name": "video", "type": "video"},
-            {"name": "participant", "type": "participant"},
-        ],
-        "params": [
-            {
-                "name": "participant",
-                "type": "participant",
-                "default": "",
-                "label": "Participant",
-                "required": True,
-            },
-        ],
-        "requires": ["videoDir"],
-    },
-    "sheet_selection": {
-        "id": "sheet_selection",
-        "label": "Sheet Selection",
-        "description": "Clip records pulled from the spreadsheet by a cell, line, or category selector.",
-        "domain": "artifact",
-        "category": "Source",
-        "inputs": [],
-        "outputs": [{"name": "clips", "type": "clipRecords"}],
-        "params": [
-            {
-                "name": "selector",
-                "type": "string",
-                "default": "",
-                "label": "Selector (cells/lines/category)",
-            },
-        ],
-        "requires": ["sheet"],
-    },
-    "time_range": {
-        "id": "time_range",
-        "label": "Time Range",
-        "description": "Manually entered start–end time ranges to drive downstream clips.",
-        "domain": "artifact",
-        "category": "Source",
-        "inputs": [],
-        "outputs": [{"name": "timeRange", "type": "timeRange"}],
-        "params": [
-            {
-                "name": "ranges",
-                "type": "string",
-                "default": "",
-                "label": "Ranges (e.g. 1:23-1:45, 2:00-2:30)",
-                "required": True,
-            },
-        ],
-        "requires": [],
-    },
-    "region": {
-        "id": "region",
-        "label": "Region",
-        "description": "A named screen region that scopes Screenspace detection to part of the frame.",
-        "domain": "screenspace",
-        "category": "Source",
-        "inputs": [],
-        "outputs": [{"name": "region", "type": "region"}],
-        "params": [
-            {"name": "name", "type": "string", "default": "", "label": "Region name"},
-        ],
-        "requires": ["videoDir"],
-    },
-    # ---- Transcript ----
-    "transcribe": {
-        "id": "transcribe",
-        "label": "Transcribe",
-        "description": "Transcribe a video's audio into a timestamped transcript and segments.",
-        "domain": "transcript",
-        "category": "Transcript",
-        "inputs": [{"name": "video", "type": "video"}],
-        "outputs": [
-            {"name": "transcript", "type": "transcript"},
-            {"name": "segments", "type": "segments"},
-        ],
-        "params": [
-            {
-                "name": "model",
-                "type": "enum",
-                "default": config.TRANSCRIBE_MODEL,
-                "choices": ["tiny", "base", "small", "medium", "large-v3"],
-                "label": "Whisper model",
-            },
-            {
-                "name": "language",
-                "type": "string",
-                "default": "auto",
-                "label": "Language",
-            },
-        ],
-        "requires": ["videoDir"],
-    },
-    "find_word": {
-        "id": "find_word",
-        "label": "Find Word",
-        "description": "Find a word or phrase in transcript segments and emit its time ranges.",
-        "domain": "transcript",
-        "category": "Transcript",
-        "inputs": [{"name": "segments", "type": "segments"}],
-        "outputs": [
-            {"name": "timeRange", "type": "timeRange"},
-            {"name": "timestamps", "type": "timestamps"},
-        ],
-        "params": [
-            {
-                "name": "word",
-                "type": "string",
-                "default": "",
-                "label": "Word or phrase",
-                "required": True,
-            },
-            {
-                "name": "pad",
-                "type": "number",
-                "default": 2,
-                "min": 0,
-                "max": 30,
-                "label": "Pad (seconds)",
-            },
-        ],
-        "requires": [],
-    },
-    "transcript_export": {
-        "id": "transcript_export",
-        "label": "Transcript Export",
-        "description": "Write a transcript file (Markdown, SRT, or VTT) from a transcript or segments.",
-        "domain": "transcript",
-        "category": "Transcript",
-        "inputs": [
-            {"name": "transcript", "type": "transcript", "optional": True},
-            {"name": "segments", "type": "segments", "optional": True},
-        ],
-        "outputs": [{"name": "artifacts", "type": "artifacts"}],
-        "params": [
-            {
-                "name": "format",
-                "type": "enum",
-                "default": config.TRANSCRIBE_FORMAT,
-                "choices": ["md", "srt", "vtt"],
-                "label": "Format",
-            },
-        ],
-        "requires": [],
-    },
-    # ---- Thinking (Ollama) ----
-    "summarize": {
-        "id": "summarize",
-        "label": "Summarize",
-        "description": "Summarize a transcript into a paragraph and key bullet points (Ollama).",
-        "domain": "thinking",
-        "category": "Thinking",
-        "inputs": [{"name": "transcript", "type": "transcript"}],
-        "outputs": [{"name": "summary", "type": "summary"}],
-        "params": [_OLLAMA_MODEL_PARAM],
-        "requires": [],
-    },
-    "citations": {
-        "id": "citations",
-        "label": "Citations",
-        "description": "Link summary claims back to the transcript segments that support them (Ollama).",
-        "domain": "thinking",
-        "category": "Thinking",
-        "inputs": [
-            {"name": "summary", "type": "summary"},
-            {"name": "segments", "type": "segments"},
-        ],
-        "outputs": [{"name": "citations", "type": "citations"}],
-        "params": [_OLLAMA_MODEL_PARAM],
-        "requires": [],
-    },
-    "friction": {
-        "id": "friction",
-        "label": "Friction",
-        "description": "Score transcript segments for usability friction and surface the rough moments.",
-        "domain": "thinking",
-        "category": "Thinking",
-        "inputs": [
-            {"name": "segments", "type": "segments"},
-            {"name": "summary", "type": "summary", "optional": True},
-        ],
-        "outputs": [{"name": "friction", "type": "friction"}],
-        "params": [_OLLAMA_MODEL_PARAM],
-        "requires": [],
-    },
-    # ---- Screenspace ----
-    # The ten per-detector nodes (ss_text … ss_boundary) are appended below the
-    # literal from ``_SS_DETECTOR_SPECS`` so each tool's real params reach the
-    # scan (the old single ``ss_scan`` passed ``parameters={}``).
-    "multitool": {
-        "id": "multitool",
-        "label": "Multitool",
-        "description": "Chain two or more per-frame detectors, matching frames where all of them fire.",
-        "domain": "screenspace",
-        "category": "Screenspace",
-        "inputs": [
-            {"name": "video", "type": "video"},
-            {"name": "region", "type": "region", "optional": True},
-        ],
-        "outputs": [{"name": "events", "type": "events"}],
-        "params": [
-            {
-                "name": "steps",
-                "type": "step-list",
-                "default": [],
-                "label": "Steps (≥2, chained per frame)",
-            },
-        ],
-        "requires": ["videoDir"],
-    },
-    "timelapse": {
-        "id": "timelapse",
-        "label": "Timelapse",
-        "description": "Condense a video into a sped-up timelapse clip or GIF.",
-        "domain": "screenspace",
-        "category": "Screenspace",
-        "inputs": [
-            {"name": "video", "type": "video"},
-            {"name": "region", "type": "region", "optional": True},
-        ],
-        "outputs": [{"name": "artifacts", "type": "artifacts"}],
-        "params": [
-            {
-                "name": "speedup_factor",
-                "type": "number",
-                "default": 10.0,
-                "min": 1,
-                "label": "Speed-up ×",
-            },
-            {
-                "name": "output_format",
-                "type": "enum",
-                "default": "mp4",
-                "choices": ["mp4", "gif"],
-                "label": "Format",
-            },
-            {
-                "name": "sample_interval",
-                "type": "number",
-                "default": 0.0,
-                "min": 0,
-                "label": "Sample interval (s)",
-            },
-        ],
-        "requires": ["videoDir"],
-    },
-    "heatmap": {
-        "id": "heatmap",
-        "label": "Heatmap",
-        "description": "Render a heatmap image from detector events (match the upstream detector style).",
-        "domain": "screenspace",
-        "category": "Screenspace",
-        "inputs": [{"name": "events", "type": "events"}],
-        "outputs": [{"name": "artifacts", "type": "artifacts"}],
-        "params": [
-            {
-                "name": "style",
-                "type": "enum",
-                "default": "change",
-                "choices": ["template", "flow", "change"],
-                "label": "Style (needs matching upstream detector)",
-            },
-            {
-                "name": "output",
-                "type": "enum",
-                "default": "image",
-                "choices": ["image", "gif", "rolling_gif"],
-                "label": "Output",
-            },
-            {
-                "name": "frames",
-                "type": "number",
-                "default": 24,
-                "min": 2,
-                "label": "GIF frames",
-            },
-            {
-                "name": "window",
-                "type": "number",
-                "default": 6,
-                "min": 1,
-                "label": "Rolling window (frames)",
-            },
-        ],
-        "requires": ["videoDir"],
-    },
-    # ---- Artifact ----
-    "highlights": {
-        "id": "highlights",
-        "label": "Highlights",
-        "description": "Score clips by severity, uniqueness, and annotations, keeping the best within a time budget.",
-        "domain": "artifact",
-        "category": "Artifact",
-        "inputs": [{"name": "clips", "type": "clipRecords"}],
-        "outputs": [{"name": "clips", "type": "clipRecords"}],
-        "params": [
-            {
-                "name": "budget",
-                "type": "number",
-                "default": config.HIGHLIGHTS_REEL_DURATION_SECONDS,
-                "min": 1,
-                "label": "Budget (seconds)",
-            },
-        ],
-        "requires": [],
-    },
-    "make_clips": {
-        "id": "make_clips",
-        "label": "Make Clips",
-        "description": "Cut clips, screenshots, or GIFs from clip records, a video, or time ranges.",
-        "domain": "artifact",
-        "category": "Artifact",
-        "inputs": [
-            {"name": "clips", "type": "clipRecords", "optional": True},
-            {"name": "video", "type": "video", "optional": True},
-            {"name": "timeRange", "type": "timeRange", "optional": True},
-        ],
-        "outputs": [{"name": "artifacts", "type": "artifacts"}],
-        "params": [
-            {
-                "name": "description",
-                "type": "string",
-                "default": "",
-                "label": "Description",
-            },
-            {
-                "name": "output_format",
-                "type": "enum",
-                "default": "clip",
-                "choices": ["clip", "screen", "gif"],
-                "label": "Output format",
-            },
-            {
-                "name": "titlecards",
-                "type": "bool",
-                "default": False,
-                "label": "Titlecards",
-            },
-            {
-                "name": "titlecard_duration",
-                "type": "number",
-                "default": config.TITLECARD_DURATION_SECONDS,
-                "min": 1,
-                "max": 30,
-                "label": "Titlecard duration (s)",
-            },
-            # Pad fields omit "min" so the number input accepts negatives
-            # (negative = trim inward); max_duration keeps min 0 (0 = no cap).
-            {
-                "name": "pad_start",
-                "type": "number",
-                "default": 0,
-                "label": "Pad start (s)",
-            },
-            {"name": "pad_end", "type": "number", "default": 0, "label": "Pad end (s)"},
-            {
-                "name": "max_duration",
-                "type": "number",
-                "default": 0,
-                "min": 0,
-                "label": "Max duration (s, 0 = none)",
-            },
-        ],
-        "requires": ["videoDir"],
-    },
-    "interval_captures": {
-        "id": "interval_captures",
-        "label": "Interval Captures",
-        "description": "Sample a screenshot or GIF at a fixed interval across the video (or each input time range).",
-        "domain": "artifact",
-        "category": "Artifact",
-        "inputs": [
-            {"name": "video", "type": "video"},
-            {"name": "timeRange", "type": "timeRange", "optional": True},
-        ],
-        "outputs": [{"name": "artifacts", "type": "artifacts"}],
-        "params": [
-            {
-                "name": "interval",
-                "type": "number",
-                "default": config.GALLERY_INTERVAL_SECONDS,
-                "min": 1,
-                "label": "Interval (s)",
-                "required": True,
-            },
-            {
-                "name": "output_format",
-                "type": "enum",
-                "default": "screen",
-                "choices": ["screen", "gif"],
-                "label": "Output format",
-            },
-            {
-                "name": "gif_duration",
-                "type": "number",
-                "default": config.GALLERY_GIF_DURATION_SECONDS,
-                "min": 1,
-                "label": "GIF duration (s)",
-            },
-        ],
-        "requires": ["videoDir"],
-    },
-    "build_reel": {
-        "id": "build_reel",
-        "label": "Build Reel",
-        "description": "Concatenate clips into a single reel video plus a manifest.",
-        "domain": "artifact",
-        "category": "Artifact",
-        "inputs": [{"name": "clips", "type": "clipRecords"}],
-        "outputs": [
-            {"name": "artifacts", "type": "artifacts"},
-            {"name": "manifest", "type": "manifest"},
-        ],
-        "params": [
-            {"name": "name", "type": "string", "default": "reel", "label": "Reel name"},
-            {
-                "name": "chronological",
-                "type": "bool",
-                "default": False,
-                "label": "Chronological order",
-            },
-            # Pad fields omit "min" so the number input accepts negatives
-            # (negative = trim inward); max_duration keeps min 0 (0 = no cap).
-            {
-                "name": "pad_start",
-                "type": "number",
-                "default": 0,
-                "label": "Pad start (s)",
-            },
-            {"name": "pad_end", "type": "number", "default": 0, "label": "Pad end (s)"},
-            {
-                "name": "max_duration",
-                "type": "number",
-                "default": 0,
-                "min": 0,
-                "label": "Max duration (s, 0 = none)",
-            },
-        ],
-        "requires": ["videoDir"],
-    },
-    "data_export": {
-        "id": "data_export",
-        "label": "Data Export",
-        "description": "Export events and segments as analysis-ready JSON/CSV tables.",
-        "domain": "artifact",
-        "category": "Artifact",
-        "inputs": [
-            {"name": "events", "type": "events", "optional": True},
-            {"name": "segments", "type": "segments", "optional": True},
-        ],
-        "outputs": [{"name": "artifacts", "type": "artifacts"}],
-        "params": [
-            {
-                "name": "format",
-                "type": "enum",
-                "default": "both",
-                "choices": ["both", "json", "csv"],
-                "label": "Format",
-            },
-        ],
-        "requires": [],
-    },
-    "timeline_viewer": {
-        "id": "timeline_viewer",
-        "label": "Timeline Viewer",
-        "description": "Bundle artifacts, events, and segments into a standalone timeline HTML viewer.",
-        "domain": "artifact",
-        "category": "Artifact",
-        "inputs": [
-            {"name": "artifacts", "type": "artifacts", "optional": True},
-            {"name": "events", "type": "events", "optional": True},
-            {"name": "segments", "type": "segments", "optional": True},
-        ],
-        "outputs": [{"name": "viewer", "type": "viewerHtml"}],
-        "params": [],
-        "requires": [],
-    },
-    # ---- Control ----
-    "measure": {
-        "id": "measure",
-        "label": "Measure",
-        "description": "Reduce events, clips, or segments to a single number (count, confidence, or duration).",
-        "domain": "control",
-        "category": "Control",
-        "inputs": [
-            {"name": "events", "type": "events", "optional": True},
-            {"name": "clips", "type": "clipRecords", "optional": True},
-            {"name": "segments", "type": "segments", "optional": True},
-        ],
-        "outputs": [{"name": "value", "type": "scalar"}],
-        "params": [
-            {
-                "name": "metric",
-                "type": "enum",
-                "default": "count",
-                "choices": ["count", "max_confidence", "total_duration"],
-                "label": "Metric",
-            },
-        ],
-        "requires": [],
-    },
-    "gate": {
-        "id": "gate",
-        "label": "Gate",
-        "description": "Compare a measured value to a threshold to allow or skip downstream nodes.",
-        "domain": "control",
-        "category": "Control",
-        "inputs": [{"name": "value", "type": "scalar"}],
-        # ``pass`` is a CONTROL output: it carries no data, it gates. The runner
-        # skips a node when an upstream gate completed with ``pass`` False, and
-        # excludes control edges from a node's data inputs. The universal
-        # ``__gate__`` input port the frontend renders is also ``control``-typed,
-        # so a gate can wire into any node (exact-match) as a control dependency.
-        "outputs": [{"name": "pass", "type": "control"}],
-        "params": [
-            {
-                "name": "op",
-                "type": "enum",
-                "default": ">=",
-                "choices": [">=", ">", "<=", "<", "==", "!="],
-                "label": "Comparison",
-            },
-            {"name": "threshold", "type": "number", "default": 0, "label": "Threshold"},
-        ],
-        "requires": [],
-    },
-    "gate_collection": {
-        "id": "gate_collection",
-        "label": "Threshold Gate",
-        "description": "Measure events, clips, or segments and gate downstream nodes on the result (Measure + Gate in one node).",
-        "domain": "control",
-        "category": "Control",
-        "inputs": [
-            {"name": "events", "type": "events", "optional": True},
-            {"name": "clips", "type": "clipRecords", "optional": True},
-            {"name": "segments", "type": "segments", "optional": True},
-        ],
-        # ``pass`` gates exactly like the plain Gate's output (see that node).
-        "outputs": [{"name": "pass", "type": "control"}],
-        "params": [
-            {
-                "name": "metric",
-                "type": "enum",
-                "default": "count",
-                "choices": ["count", "max_confidence", "total_duration"],
-                "label": "Metric",
-            },
-            {
-                "name": "op",
-                "type": "enum",
-                "default": ">=",
-                "choices": [">=", ">", "<=", "<", "==", "!="],
-                "label": "Comparison",
-            },
-            {"name": "threshold", "type": "number", "default": 0, "label": "Threshold"},
-        ],
-        "requires": [],
-    },
-}
-
-
-# Per-detector Screenspace nodes. Each entry's params are lifted from the matching
-# ``screenspace_tools`` class (the knobs its ``scan`` reads). The three
-# reference-based detectors (template/similarity/scene) self-extract their
-# reference from the node's region at ``reference_seconds`` so the canvas needs no
-# upload UI. ``_build_ss_scan_params`` (below) assembles these flat params into the
-# nested ``scan_params`` each scan expects.
-_SS_DETECTOR_LABELS: dict[str, str] = {
-    "text": "Detect Text",
-    "color": "Detect Color",
-    "change": "Detect Change",
-    "similarity": "Detect Similarity",
-    "numbers": "Detect Numbers",
-    "template": "Detect Template",
-    "flow": "Detect Motion",
-    "scene": "Detect Scene",
-    "inactivity": "Detect Inactivity",
-    "boundary": "Detect Boundary",
-}
-
-_SS_DETECTOR_DESCRIPTIONS: dict[str, str] = {
-    "text": "Detect when target text appears in the region via OCR.",
-    "color": "Detect when a target color appears or covers enough of the region.",
-    "change": "Detect visual change between frames in the region.",
-    "similarity": "Detect frames matching a reference image sampled from the region.",
-    "numbers": "Read numbers in the region via OCR and compare them to a target.",
-    "template": "Detect a reference template (sampled from the region) appearing in the frame.",
-    "flow": "Detect motion in the region via optical flow.",
-    "scene": "Detect scene changes against a reference fingerprint sampled from the region.",
-    "inactivity": "Detect stretches of inactivity (no change) in the region.",
-    "boundary": "Detect UI boundaries or edges appearing in the region.",
-}
-
-_INTERVAL_PARAM: ParamSpec = {
-    "name": "interval",
-    "type": "number",
-    "default": 0,
-    "min": 0,
-    "label": "Interval (s, 0=auto)",
-}
-
-_SS_DETECTOR_SPECS: dict[str, list[ParamSpec]] = {
-    "text": [
-        {
-            "name": "search_string",
-            "type": "string",
-            "default": "",
-            "label": "Search text",
-            "required": True,
-        },
-        {
-            "name": "fuzzy_threshold",
-            "type": "number",
-            "default": 80,
-            "min": 0,
-            "max": 100,
-            "label": "Fuzzy match %",
-        },
-        {
-            "name": "interval",
-            "type": "number",
-            "default": 2.0,
-            "min": 0,
-            "label": "Interval (s, 0=auto)",
-        },
-    ],
-    "color": [
-        {
-            "name": "color_h",
-            "type": "number",
-            "default": 0,
-            "min": 0,
-            "max": 179,
-            "label": "Hue",
-        },
-        {
-            "name": "color_s",
-            "type": "number",
-            "default": 0,
-            "min": 0,
-            "max": 255,
-            "label": "Saturation",
-        },
-        {
-            "name": "color_v",
-            "type": "number",
-            "default": 0,
-            "min": 0,
-            "max": 255,
-            "label": "Value",
-        },
-        {
-            "name": "tol_h",
-            "type": "number",
-            "default": 10,
-            "min": 0,
-            "max": 179,
-            "label": "Hue tol",
-        },
-        {
-            "name": "tol_s",
-            "type": "number",
-            "default": 50,
-            "min": 0,
-            "max": 255,
-            "label": "Sat tol",
-        },
-        {
-            "name": "tol_v",
-            "type": "number",
-            "default": 50,
-            "min": 0,
-            "max": 255,
-            "label": "Val tol",
-        },
-        {
-            "name": "color_mode",
-            "type": "enum",
-            "default": "average",
-            "choices": ["average", "presence"],
-            "label": "Mode",
-        },
-        {
-            "name": "min_coverage",
-            "type": "number",
-            "default": 0.0,
-            "min": 0,
-            "max": 1,
-            "label": "Min coverage",
-        },
-        _INTERVAL_PARAM,
-    ],
-    "change": [
-        {
-            "name": "threshold",
-            "type": "number",
-            "default": config.SCREENSPACE_CHANGE_RATIO_THRESHOLD,
-            "min": 0,
-            "max": 1,
-            "label": "Change ratio",
-        },
-        {
-            "name": "noise_threshold",
-            "type": "number",
-            "default": config.SCREENSPACE_NOISE_THRESHOLD,
-            "min": 0,
-            "label": "Noise threshold",
-        },
-        {
-            "name": "require_consecutive",
-            "type": "number",
-            "default": 1,
-            "min": 1,
-            "label": "Consecutive frames",
-        },
-        _INTERVAL_PARAM,
-    ],
-    "similarity": [
-        {
-            "name": "reference_seconds",
-            "type": "number",
-            "default": 0.0,
-            "min": 0,
-            "label": "Reference time (s)",
-        },
-        {
-            "name": "threshold",
-            "type": "number",
-            "default": config.SCREENSPACE_SSIM_THRESHOLD,
-            "min": 0,
-            "max": 1,
-            "label": "SSIM threshold",
-        },
-        _INTERVAL_PARAM,
-    ],
-    "numbers": [
-        {
-            "name": "operator",
-            "type": "enum",
-            "default": "gt",
-            "choices": ["gt", "lt", "gte", "lte", "eq", "range"],
-            "label": "Operator",
-        },
-        {
-            "name": "target_value",
-            "type": "number",
-            "default": 0,
-            "label": "Target value",
-        },
-        {"name": "range_min", "type": "number", "default": 0, "label": "Range min"},
-        {"name": "range_max", "type": "number", "default": 0, "label": "Range max"},
-        {
-            "name": "integers_only",
-            "type": "bool",
-            "default": False,
-            "label": "Integers only",
-        },
-        {
-            "name": "interval",
-            "type": "number",
-            "default": 2.0,
-            "min": 0,
-            "label": "Interval (s, 0=auto)",
-        },
-    ],
-    "template": [
-        {
-            "name": "reference_seconds",
-            "type": "number",
-            "default": 0.0,
-            "min": 0,
-            "label": "Reference time (s)",
-        },
-        {
-            "name": "threshold",
-            "type": "number",
-            "default": config.SCREENSPACE_TEMPLATE_MATCH_THRESHOLD,
-            "min": 0,
-            "max": 1,
-            "label": "Match threshold",
-        },
-        {
-            "name": "template_scale",
-            "type": "number",
-            "default": 1.0,
-            "min": 0,
-            "label": "Template scale",
-        },
-        _INTERVAL_PARAM,
-    ],
-    "flow": [
-        {
-            "name": "magnitude_threshold",
-            "type": "number",
-            "default": config.SCREENSPACE_FLOW_MAGNITUDE_THRESHOLD,
-            "min": 0,
-            "label": "Magnitude threshold",
-        },
-        {
-            "name": "require_consecutive",
-            "type": "number",
-            "default": 1,
-            "min": 1,
-            "label": "Consecutive frames",
-        },
-        _INTERVAL_PARAM,
-    ],
-    "scene": [
-        {
-            "name": "reference_seconds",
-            "type": "number",
-            "default": 0.0,
-            "min": 0,
-            "label": "Reference time (s)",
-        },
-        {
-            "name": "threshold",
-            "type": "number",
-            "default": config.SCREENSPACE_SCENE_SIMILARITY_THRESHOLD,
-            "min": 0,
-            "max": 1,
-            "label": "Scene threshold",
-        },
-        _INTERVAL_PARAM,
-    ],
-    "inactivity": [
-        {
-            "name": "threshold",
-            "type": "number",
-            "default": config.SCREENSPACE_INACTIVITY_PHASH_THRESHOLD,
-            "min": 0,
-            "label": "Phash threshold",
-        },
-        {
-            "name": "min_duration",
-            "type": "number",
-            "default": 0.0,
-            "min": 0,
-            "label": "Min duration (s)",
-        },
-        _INTERVAL_PARAM,
-    ],
-    "boundary": [
-        {
-            "name": "threshold",
-            "type": "number",
-            "default": 0,
-            "min": 0,
-            "label": "Threshold (0=auto)",
-        },
-        {
-            "name": "min_gap",
-            "type": "number",
-            "default": 0.0,
-            "min": 0,
-            "label": "Min gap (s)",
-        },
-        {
-            "name": "metric",
-            "type": "enum",
-            "default": config.SCREENSPACE_BOUNDARY_METRIC,
-            "choices": ["hybrid", "phash", "scene"],
-            "label": "Metric",
-        },
-        _INTERVAL_PARAM,
-    ],
-}
-
-# Detectors whose scan needs a reference frame self-extracted from the node region.
-_SS_REFERENCE_DETECTORS = frozenset({"similarity", "template", "scene"})
-
-# Detectors usable as a Multitool chain step: the per-frame (``check_frame``)
-# detectors that need no uploaded reference. Single source of truth — served to
-# the frontend via each node's ``multitoolStep`` flag so the step editor derives
-# the list instead of hardcoding it (the "no duplicated JS constants" rule).
-# ``tests/test_workflows_executors`` cross-checks this against the actual tool
-# classes (override ``check_frame`` AND not reference-based) so it can't drift.
-_MULTITOOL_STEP_TOOLS = frozenset(
-    {"color", "change", "flow", "text", "numbers", "inactivity"}
-)
-
-for _ss_tool in _SS_DETECTOR_SPECS:
-    NODE_TYPES[f"ss_{_ss_tool}"] = {
-        "id": f"ss_{_ss_tool}",
-        "label": _SS_DETECTOR_LABELS[_ss_tool],
-        "description": _SS_DETECTOR_DESCRIPTIONS[_ss_tool],
-        "domain": "screenspace",
-        "category": "Screenspace",
-        "inputs": [
-            {"name": "video", "type": "video"},
-            {"name": "region", "type": "region", "optional": True},
-            {"name": "timeRange", "type": "timeRange", "optional": True},
-        ],
-        "outputs": [{"name": "events", "type": "events"}],
-        "params": list(_SS_DETECTOR_SPECS[_ss_tool]),
-        "requires": ["videoDir"],
-        # Hidden from the palette: the unified "detect" node below is the
-        # palette-facing entry. These stay in the catalog as the per-detector
-        # spec source (Detect editor + Multitool steps) and keep old blueprints
-        # and built-in recipes that reference ss_<tool> directly runnable.
-        "hidden": True,
-        # Whether this detector can be a Multitool step (the frontend derives the
-        # step-type list from this flag — see _MULTITOOL_STEP_TOOLS).
-        "multitoolStep": _ss_tool in _MULTITOOL_STEP_TOOLS,
-    }
-
-# Unified palette-facing detector: one node whose ``detector`` dropdown swaps the
-# per-detector param set (the frontend reads it from the hidden ss_<tool> nodes
-# above). Dispatches to the same _run_ss_detector body the ss_<tool> nodes use.
-NODE_TYPES["detect"] = {
-    "id": "detect",
-    "label": "Detect",
-    "description": "Detect a region condition (text, colour, motion, numbers). Pick the detector.",
-    "domain": "screenspace",
-    "category": "Screenspace",
-    "inputs": [
-        {"name": "video", "type": "video"},
-        {"name": "region", "type": "region", "optional": True},
-        {"name": "timeRange", "type": "timeRange", "optional": True},
-    ],
-    "outputs": [{"name": "events", "type": "events"}],
-    "params": [
-        {
-            "name": "detector",
-            "type": "enum",
-            "default": "text",
-            "choices": list(_SS_DETECTOR_SPECS.keys()),
-            "label": "Detector",
-        }
-    ],
-    "requires": ["videoDir"],
-}
-
-
-def serialize_catalog() -> list[dict[str, Any]]:
-    """Return ``NODE_TYPES`` as ordered JSON-safe dicts (without ``execute``).
-
-    Drives ``GET /workflows/api/catalog``. Deliberately *not* routed through
-    ``utils.get_frontend_config`` — the catalog is large and Workflows-specific,
-    and bolting it onto the shared config would pollute the
-    ``tests/test_shared_constants`` contract.
-    """
-    return [
-        {k: v for k, v in node.items() if k != "execute"}
-        for node in NODE_TYPES.values()
-    ]
-
-
-def serialize_adapters() -> list[list[str]]:
-    """Return the ``ADAPTERS`` keys as JSON-safe ``[src, dst, description]`` rows.
-
-    Drives the ``adapters`` field of ``GET /workflows/api/catalog`` so the
-    frontend's ``canConnect`` can accept the same coercions the runner applies
-    (``_gather_inputs``), and so a coerced wire's tooltip can explain the
-    transformation. Serving the table — rather than duplicating it in JS — keeps
-    UI wire-validity in lockstep with the runner (``ADAPTERS`` defined below is the
-    single source of truth; ``tests/test_workflows_api`` guards parity).
-    """
-    return [
-        [src, dst, _ADAPTER_DESCRIPTIONS.get((src, dst), "")] for src, dst in ADAPTERS
-    ]
-
-
-# ---------------------------------------------------------------------------
-# Built-in recipes (P4) — read-only stashes served alongside user stashes
-# ---------------------------------------------------------------------------
-#
-# These are the headline graphs shipped as ready-to-stamp sub-graphs. They are
-# *code, not data*: ``GET /api/stashes`` prepends them to the user's persisted
-# stashes, so they are never seeded into the manifest (no migration/dedup). The
-# ``builtin`` flag makes them read-only — the stash CRUD routes reject renaming
-# or deleting them. Node/edge shapes mirror the on-canvas blueprint shapes
-# (``{id, type, params, position}`` / ``{id, from, fromPort, to, toPort}``) with
-# stash-local ids; the frontend remaps to fresh ``n_``/``e_`` ids on instantiate.
-
-BUILTIN_STASHES: list[dict[str, Any]] = [
-    {
-        "id": "builtin_clip_word",
-        "name": "Transcribe → Find Word → Make Clips → Viewer",
-        "builtin": True,
-        "createdAt": "",
-        "nodes": [
-            {
-                "id": "s1",
-                "type": "video_source",
-                "params": {"participant": ""},
-                "position": {"x": 40, "y": 160},
-            },
-            {
-                "id": "s2",
-                "type": "transcribe",
-                "params": {"language": "auto"},
-                "position": {"x": 340, "y": 80},
-            },
-            {
-                "id": "s3",
-                "type": "find_word",
-                "params": {"word": "", "pad": 2},
-                "position": {"x": 640, "y": 80},
-            },
-            {
-                "id": "s4",
-                "type": "make_clips",
-                "params": {
-                    "description": "",
-                    "output_format": "clip",
-                    "titlecards": False,
-                    "titlecard_duration": config.TITLECARD_DURATION_SECONDS,
-                },
-                "position": {"x": 940, "y": 160},
-            },
-            {
-                "id": "s5",
-                "type": "timeline_viewer",
-                "params": {},
-                "position": {"x": 1240, "y": 160},
-            },
-        ],
-        "edges": [
-            {
-                "id": "se1",
-                "from": "s1",
-                "fromPort": "video",
-                "to": "s2",
-                "toPort": "video",
-            },
-            {
-                "id": "se2",
-                "from": "s2",
-                "fromPort": "segments",
-                "to": "s3",
-                "toPort": "segments",
-            },
-            {
-                "id": "se3",
-                "from": "s3",
-                "fromPort": "timeRange",
-                "to": "s4",
-                "toPort": "timeRange",
-            },
-            {
-                "id": "se4",
-                "from": "s1",
-                "fromPort": "video",
-                "to": "s4",
-                "toPort": "video",
-            },
-            {
-                "id": "se5",
-                "from": "s4",
-                "fromPort": "artifacts",
-                "to": "s5",
-                "toPort": "artifacts",
-            },
-        ],
-    },
-    {
-        "id": "builtin_highlights_reel",
-        "name": "Sheet Selection → Highlights → Build Reel → Viewer",
-        "builtin": True,
-        "createdAt": "",
-        "nodes": [
-            {
-                "id": "s1",
-                "type": "sheet_selection",
-                "params": {"selector": ""},
-                "position": {"x": 40, "y": 160},
-            },
-            {
-                "id": "s2",
-                "type": "highlights",
-                "params": {"budget": config.HIGHLIGHTS_REEL_DURATION_SECONDS},
-                "position": {"x": 340, "y": 160},
-            },
-            {
-                "id": "s3",
-                "type": "build_reel",
-                "params": {"name": "reel"},
-                "position": {"x": 640, "y": 160},
-            },
-            {
-                "id": "s4",
-                "type": "timeline_viewer",
-                "params": {},
-                "position": {"x": 940, "y": 160},
-            },
-        ],
-        "edges": [
-            {
-                "id": "se1",
-                "from": "s1",
-                "fromPort": "clips",
-                "to": "s2",
-                "toPort": "clips",
-            },
-            {
-                "id": "se2",
-                "from": "s2",
-                "fromPort": "clips",
-                "to": "s3",
-                "toPort": "clips",
-            },
-            {
-                "id": "se3",
-                "from": "s3",
-                "fromPort": "artifacts",
-                "to": "s4",
-                "toPort": "artifacts",
-            },
-        ],
-    },
-    {
-        # Demonstrates the collection-algebra family (limit_events): keep only the
-        # top-N most confident detections before cutting clips.
-        "id": "builtin_top_detections",
-        "name": "Detect → Top Events → Make Clips → Viewer",
-        "builtin": True,
-        "createdAt": "",
-        "nodes": [
-            {
-                "id": "s1",
-                "type": "video_source",
-                "params": {"participant": ""},
-                "position": {"x": 40, "y": 160},
-            },
-            {
-                "id": "s2",
-                "type": "detect",
-                "params": {"detector": "text"},
-                "position": {"x": 340, "y": 80},
-            },
-            {
-                "id": "s3",
-                "type": "limit_events",
-                "params": {"sort_by": "confidence", "order": "desc", "take": 10},
-                "position": {"x": 640, "y": 80},
-            },
-            {
-                "id": "s4",
-                "type": "make_clips",
-                "params": {
-                    "description": "",
-                    "output_format": "clip",
-                    "titlecards": False,
-                    "titlecard_duration": config.TITLECARD_DURATION_SECONDS,
-                },
-                "position": {"x": 940, "y": 160},
-            },
-            {
-                "id": "s5",
-                "type": "timeline_viewer",
-                "params": {},
-                "position": {"x": 1240, "y": 160},
-            },
-        ],
-        "edges": [
-            {
-                "id": "se1",
-                "from": "s1",
-                "fromPort": "video",
-                "to": "s2",
-                "toPort": "video",
-            },
-            {
-                "id": "se2",
-                "from": "s2",
-                "fromPort": "events",
-                "to": "s3",
-                "toPort": "in",
-            },
-            {
-                "id": "se3",
-                "from": "s3",
-                "fromPort": "out",
-                "to": "s4",
-                "toPort": "clips",
-            },
-            {
-                "id": "se4",
-                "from": "s1",
-                "fromPort": "video",
-                "to": "s4",
-                "toPort": "video",
-            },
-            {
-                "id": "se5",
-                "from": "s4",
-                "fromPort": "artifacts",
-                "to": "s5",
-                "toPort": "artifacts",
-            },
-        ],
-    },
-    {
-        "id": "builtin_transcript_exports",
-        "name": "Transcribe → Transcript + Data Export",
-        "builtin": True,
-        "createdAt": "",
-        "nodes": [
-            {
-                "id": "s1",
-                "type": "video_source",
-                "params": {"participant": ""},
-                "position": {"x": 40, "y": 160},
-            },
-            {
-                "id": "s2",
-                "type": "transcribe",
-                "params": {"language": "auto"},
-                "position": {"x": 340, "y": 160},
-            },
-            {
-                "id": "s3",
-                "type": "transcript_export",
-                "params": {"format": config.TRANSCRIBE_FORMAT},
-                "position": {"x": 640, "y": 80},
-            },
-            {
-                "id": "s4",
-                "type": "data_export",
-                "params": {"format": "both"},
-                "position": {"x": 640, "y": 240},
-            },
-        ],
-        "edges": [
-            {
-                "id": "se1",
-                "from": "s1",
-                "fromPort": "video",
-                "to": "s2",
-                "toPort": "video",
-            },
-            {
-                "id": "se2",
-                "from": "s2",
-                "fromPort": "transcript",
-                "to": "s3",
-                "toPort": "transcript",
-            },
-            {
-                "id": "se3",
-                "from": "s2",
-                "fromPort": "segments",
-                "to": "s4",
-                "toPort": "segments",
-            },
-        ],
-    },
-]
-
-
-# ---------------------------------------------------------------------------
 # Executors (M3) — thin adapters over existing pure functions
 # ---------------------------------------------------------------------------
 #
@@ -1512,53 +196,6 @@ BUILTIN_STASHES: list[dict[str, Any]] = [
 # carried on each wire is documented in ``plans/archive/WORKFLOWS-PLAN.md``; the unifying
 # primitive is a "source descriptor" embedded in every domain value so the pure
 # ``ADAPTERS`` (value -> value, no ctx/params) can still reach a clip's source.
-
-_DEFAULT_EVENT_CLUSTER_GAP = 5.0  # seconds; matches the CLI --cluster-gap default
-
-
-def _study_from_filename(filename: str) -> str:
-    """Derive the study name from a ``{study}_{pid}`` basename ('' when absent)."""
-    head, _sep, _tail = Path(filename).stem.rpartition("_")
-    return head
-
-
-def _source_descriptor(participant: str, video_paths: list[str]) -> dict[str, Any]:
-    """Build the source descriptor embedded in every domain value."""
-    first = Path(video_paths[0]).name if video_paths else ""
-    return {
-        "participant": participant,
-        "study": _study_from_filename(first),
-        "source_filename": first,
-        "video_paths": list(video_paths),
-    }
-
-
-def _source_from_events(events: list[dict[str, Any]]) -> dict[str, Any]:
-    """Best-effort source descriptor derived from a screenspace event list."""
-    if not events:
-        return {}
-    first = events[0]
-    source_video = str(first.get("source_video", "") or "")
-    return {
-        "participant": str(first.get("participant", "") or ""),
-        "study": _study_from_filename(source_video),
-        "source_filename": source_video,
-        "video_paths": [source_video] if source_video else [],
-    }
-
-
-def _clip_source_filename(source: dict[str, Any]) -> str:
-    """Source-video override for clip cutting (``+``-joined for multi-part).
-
-    ``pipeline._check_source_video`` resolves a ``" + "``-joined override into all
-    parts (building the clip's ``source_timeline``); a single basename resolves by
-    exact match. So a multi-video participant flows through clip cutting with the
-    right global-time mapping without any timeline code here.
-    """
-    paths = source.get("video_paths") or []
-    if len(paths) >= 2:
-        return " + ".join(Path(p).name for p in paths)
-    return str(source.get("source_filename", "") or "")
 
 
 # ---- Sources ----
@@ -3264,151 +1901,6 @@ def _limit_params(kind: str) -> list[ParamSpec]:
     ]
 
 
-# ---------------------------------------------------------------------------
-# Typed-port adapters (M3) — pure value -> value, applied by the runner (M4)
-# ---------------------------------------------------------------------------
-#
-# An adapter coerces an output value whose port type differs from the consuming
-# input's type. They are pure (no ctx/params), so each value self-carries its
-# source descriptor (see executors above).
-
-
-def _adapt_transcript_to_segments(value: dict[str, Any]) -> dict[str, Any]:
-    return {
-        "segments": value.get("segments", []),
-        "source": value.get("source", {}),
-    }
-
-
-def _adapt_segments_to_timerange(value: dict[str, Any]) -> dict[str, Any]:
-    """Project every segment to a ``(start, end)`` window (no text filter).
-
-    Distinct from the ``find_word`` node, which filters segments by a search term;
-    this is the default coercion when ``segments`` is wired straight into a
-    ``timeRange`` input.
-    """
-    ranges: list[tuple[float, float]] = []
-    for seg in value.get("segments") or []:
-        start = float(seg.get("start", 0.0) or 0.0)
-        end = float(seg.get("end", start) or start)
-        ranges.append((start, max(start, end)))
-    return {"ranges": ranges, "source": value.get("source", {})}
-
-
-def _adapt_timerange_to_cliprecords(value: dict[str, Any]) -> dict[str, Any]:
-    import files
-
-    source = value.get("source") or {}
-    ranges = [(float(s), float(e)) for s, e in (value.get("ranges") or [])]
-    records = files.build_clip_records(
-        participant=str(source.get("participant", "") or ""),
-        source_filename=_clip_source_filename(source),
-        time_ranges=ranges,
-        description="workflow",
-        study=str(source.get("study", "") or ""),
-    )
-    return {"records": records, "study": str(source.get("study", "") or "")}
-
-
-def _adapt_events_to_timerange(value: dict[str, Any]) -> dict[str, Any]:
-    events = value.get("events") or []
-    ranges: list[tuple[float, float]] = []
-    for ev in events:
-        t_in = float(ev.get("time_in", 0.0) or 0.0)
-        t_out = float(ev.get("time_out", t_in) or t_in)
-        ranges.append((t_in, max(t_in, t_out)))
-    source = value.get("source") or _source_from_events(events)
-    return {"ranges": ranges, "source": source}
-
-
-def _adapt_cliprecords_to_timerange(value: dict[str, Any]) -> dict[str, Any]:
-    """Project clip records to ``(start, end)`` windows — e.g. *sheet cells →
-    SS scan windows*. Each record's resolved ``times`` (via ``files.prepare_clip``
-    when not pre-filled) are converted to seconds; the inverse of
-    :func:`_adapt_timerange_to_cliprecords`.
-    """
-    import files
-
-    ranges: list[tuple[float, float]] = []
-    source: dict[str, Any] = {}
-    for rec in value.get("records") or []:
-        prepared = (
-            rec
-            if rec.get("times")
-            else files.prepare_clip(cast(utils.ClipRecord, dict(rec)))
-        )
-        for start_str, end_str in prepared.get("times") or []:
-            start = utils.timestamp_to_seconds(start_str)
-            end = utils.timestamp_to_seconds(end_str)
-            if start is not None and end is not None:
-                ranges.append((start, max(start, end)))
-        if not source and rec.get("participant"):
-            source = {
-                "participant": str(rec.get("participant", "") or ""),
-                "study": str(rec.get("study", value.get("study", "")) or ""),
-                "source_filename": "",
-                "video_paths": [],
-            }
-    return {"ranges": ranges, "source": source}
-
-
-def _adapt_events_to_cliprecords(value: dict[str, Any]) -> dict[str, Any]:
-    import files
-
-    events = value.get("events") or []
-    source = value.get("source") or _source_from_events(events)
-    spans: list[tuple[float, float]] = []
-    for ev in events:
-        t_in = float(ev.get("time_in", 0.0) or 0.0)
-        t_out = float(ev.get("time_out", t_in) or t_in)
-        spans.append((t_in, max(t_in, t_out)))
-    records = files.build_clip_records(
-        participant=str(source.get("participant", "") or ""),
-        source_filename=_clip_source_filename(source),
-        time_ranges=spans,
-        description="event",
-        category="workflow",
-        study=str(source.get("study", "") or ""),
-        cluster_gap=_DEFAULT_EVENT_CLUSTER_GAP,
-    )
-    return {"records": records, "study": str(source.get("study", "") or "")}
-
-
-def _adapt_video_to_scalar(value: dict[str, Any]) -> float:
-    import video
-
-    paths = value.get("video_paths") or []
-    if not paths:
-        return 0.0
-    props = video.probe_video_properties(paths[0]) or {}
-    return float(props.get("duration", 0.0) or 0.0)
-
-
-ADAPTERS: dict[tuple[str, str], Callable[[Any], Any]] = {
-    ("transcript", "segments"): _adapt_transcript_to_segments,
-    ("segments", "timeRange"): _adapt_segments_to_timerange,
-    ("timeRange", "clipRecords"): _adapt_timerange_to_cliprecords,
-    ("clipRecords", "timeRange"): _adapt_cliprecords_to_timerange,
-    ("events", "timeRange"): _adapt_events_to_timerange,
-    ("events", "clipRecords"): _adapt_events_to_cliprecords,
-    ("video", "scalar"): _adapt_video_to_scalar,
-}
-
-# Plain-language description of what each adapter does, served alongside the
-# table (see ``serialize_adapters``) so a coerced (dashed) wire's tooltip can
-# explain the transformation — not just that one happened. One per ADAPTERS key
-# (guarded by ``tests/test_workflows_api``); a missing one degrades to no suffix.
-_ADAPTER_DESCRIPTIONS: dict[tuple[str, str], str] = {
-    ("transcript", "segments"): "use the transcript's segments",
-    ("segments", "timeRange"): "use each segment's time span",
-    ("timeRange", "clipRecords"): "make a clip from each time range",
-    ("clipRecords", "timeRange"): "use each clip's time span",
-    ("events", "timeRange"): "use each event's time span",
-    ("events", "clipRecords"): "make a clip from each event (clustered)",
-    ("video", "scalar"): "use the video's duration",
-}
-
-
 # Wire each declarative NodeType to its executor. ``serialize_catalog`` strips
 # ``execute`` again for the JSON catalog endpoint, so this stays server-internal.
 _EXECUTORS: dict[
@@ -3536,803 +2028,3 @@ for _kind in ("events", "clips", "timerange"):
 
 for _node_id, _executor in _EXECUTORS.items():
     NODE_TYPES[_node_id]["execute"] = _executor
-
-
-# ---------------------------------------------------------------------------
-# Run engine (M4) — DAG topo-sort + sequential ready-set execution
-# ---------------------------------------------------------------------------
-#
-# ``WorkflowRunner`` runs one blueprint on a daemon thread (the server spawns it).
-# It calls the executors directly with the uniform ``NodeContext`` contract, so a
-# cross-domain DAG gets clean end-to-end progress + cancellation without routing
-# through the per-domain worker queues. Execution is strictly sequential (the v1
-# decision: Whisper/Ollama are single-resource); intra-node parallelism (e.g.
-# ``process_clips``' own pool) still applies.
-
-# Run + per-node status constants (mirrors screenspace_manifest's TASK_STATUS_*).
-RUN_STATUS_QUEUED = "queued"
-RUN_STATUS_RUNNING = "running"
-RUN_STATUS_COMPLETED = "completed"
-RUN_STATUS_FAILED = "failed"
-RUN_STATUS_CANCELLED = "cancelled"
-
-NODE_STATUS_QUEUED = "queued"
-NODE_STATUS_RUNNING = "running"
-NODE_STATUS_COMPLETED = "completed"
-NODE_STATUS_FAILED = "failed"
-NODE_STATUS_SKIPPED = "skipped"
-
-# Canvas-only sticky-note pseudo-node (frontend-created, not in NODE_TYPES).
-# Notes live in blueprint["nodes"] so they ride save/undo/copy/import for free;
-# the runner filters them out so they never execute or appear in run snapshots.
-NOTE_NODE_TYPE = "note"
-
-# Auto-run trigger types: new_video (the original watch-dir P6 trigger) plus
-# the chaining triggers (a transcript or Screenspace scan completing fires an
-# armed blueprint for that participant). Served through /api/catalog context so
-# the frontend picker never duplicates the list; workflows_server's watcher
-# polls each type's source (input dir / transcripts manifest / screenspace
-# manifest) only while a blueprint of that type is armed.
-TRIGGER_TYPES: list[dict[str, str]] = [
-    {"id": "new_video", "label": "New video lands"},
-    {"id": "transcript_complete", "label": "Transcript completes"},
-    {"id": "scan_event", "label": "Screenspace scan completes"},
-]
-TRIGGER_TYPE_IDS = frozenset(t["id"] for t in TRIGGER_TYPES)
-
-_PROGRESS_NOTIFY_INTERVAL = (
-    0.5  # seconds; throttle SSE notifies (copy screenspace_worker)
-)
-
-
-class WorkflowCycleError(ValueError):
-    """Raised by :func:`topo_order` when the graph is not a DAG (rejected at submit)."""
-
-
-def _now_iso() -> str:
-    """UTC ISO-8601 timestamp for run/node start+complete stamps."""
-    return datetime.now(timezone.utc).isoformat()
-
-
-def topo_order(nodes: list[dict[str, Any]], edges: list[dict[str, Any]]) -> list[str]:
-    """Return node ids in execution order (Kahn); raise on a cycle.
-
-    Stable: ties break by the given node order, so a run is reproducible. Edges
-    referencing unknown node ids are ignored (a stale wire never blocks a run).
-    """
-    ids = [n["id"] for n in nodes]
-    id_set = set(ids)
-    adj: dict[str, list[str]] = {nid: [] for nid in ids}
-    indeg: dict[str, int] = {nid: 0 for nid in ids}
-    for edge in edges:
-        src, dst = edge.get("from"), edge.get("to")
-        if src in id_set and dst in id_set:
-            adj[src].append(dst)
-            indeg[dst] += 1
-    ready = [nid for nid in ids if indeg[nid] == 0]
-    order: list[str] = []
-    while ready:
-        nid = ready.pop(0)
-        order.append(nid)
-        for nxt in adj[nid]:
-            indeg[nxt] -= 1
-            if indeg[nxt] == 0:
-                ready.append(nxt)
-    if len(order) != len(ids):
-        raise WorkflowCycleError("Workflow graph contains a cycle")
-    return order
-
-
-def blueprint_participant_nodes(blueprint: dict[str, Any]) -> list[dict[str, Any]]:
-    """The blueprint's participant-bound source nodes (``video_source``).
-
-    Whole-study batch (P3) fans out over these — a blueprint with none can't be
-    rebound per participant, so the batch endpoint rejects it.
-    """
-    return [n for n in blueprint.get("nodes", []) if n.get("type") == "video_source"]
-
-
-def bind_participant(blueprint: dict[str, Any], participant: str) -> dict[str, Any]:
-    """Deep-copy ``blueprint`` and rebind every ``video_source`` node to ``participant``.
-
-    Pure: the original is never mutated. Non-source nodes are untouched — only the
-    participant-bound sources are rewritten, so one blueprint can run once per
-    participant in a batch (P3).
-    """
-    clone = copy.deepcopy(blueprint)
-    for node in clone.get("nodes", []):
-        if node.get("type") == "video_source":
-            params = node.setdefault("params", {})
-            params["participant"] = participant
-    return clone
-
-
-def _port_type(type_id: str, port_name: str | None, direction: str) -> str | None:
-    """Look up a port's wire type from ``NODE_TYPES`` ('in' or 'out' direction)."""
-    node_type = NODE_TYPES.get(type_id)
-    if not node_type or not port_name:
-        return None
-    ports = node_type["outputs"] if direction == "out" else node_type["inputs"]
-    for port in ports:
-        if port["name"] == port_name:
-            return port["type"]
-    return None
-
-
-def _port_optional(type_id: str | None, port_name: str | None) -> bool:
-    """True if a node type's *input* port is declared ``optional`` (else False)."""
-    node_type = NODE_TYPES.get(type_id) if type_id else None
-    if not node_type or not port_name:
-        return False
-    for port in node_type["inputs"]:
-        if port["name"] == port_name:
-            return bool(port.get("optional"))
-    return False
-
-
-def _summarize_value(value: Any) -> Any:
-    """Shrink a node output to a JSON-safe summary (counts + terminal pointers).
-
-    The full result (raw frames, whole segment lists, ClipRecords) stays in the
-    runner's in-memory store; the snapshot ships only what a panel needs.
-    """
-    if value is None or isinstance(value, (str, int, float, bool)):
-        return value
-    if isinstance(value, list):
-        return {"count": len(value)}
-    if isinstance(value, dict):
-        out: dict[str, Any] = {}
-        for list_key in (
-            "events",
-            "segments",
-            "artifacts",
-            "records",
-            "ranges",
-            "times",
-        ):
-            seq = value.get(list_key)
-            if isinstance(seq, list):
-                out["count"] = len(seq)
-                break
-        for scalar_key in ("path", "study", "name", "participant", "count"):
-            sval = value.get(scalar_key)
-            if isinstance(sval, (str, int, float, bool)):
-                out[scalar_key] = sval
-        return out
-    return type(value).__name__
-
-
-def _node_result_summary(result: Any) -> dict[str, Any]:
-    """Per-port summary of a completed node's output (see :func:`_summarize_value`)."""
-    if not isinstance(result, dict):
-        return {}
-    return {port: _summarize_value(val) for port, val in result.items()}
-
-
-# ---- Per-node result sidecars (P5) ----------------------------------------
-#
-# The snapshot ships only counts/pointers; the *full* inspectable result is
-# written to ``<output_dir>/workflow_runs/<run_id>/<node_id>.json`` so the
-# run-history UI can lazily fetch and render it after the runner is evicted.
-
-# Output port types the run-history UI renders on row-expand. Plumbing types
-# (video/region/timeRange handles, control) are persisted for resume (below)
-# but hidden from the inspector.
-_INSPECTABLE_PORT_TYPES = frozenset(
-    {
-        "artifacts",
-        "events",
-        "segments",
-        "summary",
-        "citations",
-        "friction",
-        "manifest",
-        "viewerHtml",
-        "scalar",
-    }
-)
-
-# Output port types persisted in the sidecar — everything JSON-safe, so a
-# later resume (``compute_resume_plan``) can reload a completed node's outputs
-# verbatim. Only ``clipRecords`` is excluded: its records carry gspread
-# ``Cell`` objects that don't survive JSON, so clipRecords producers always
-# re-run on resume (cheap — one Sheets read).
-_SIDECAR_PORT_TYPES = _INSPECTABLE_PORT_TYPES | frozenset(
-    {
-        "transcript",
-        "timeRange",
-        "timestamps",
-        "video",
-        "participant",
-        "region",
-        "control",  # a gate's pass verdict is a JSON bool; resume needs it
-    }
-)
-
-
-def run_results_dir(output_dir: Path | str, run_id: str) -> Path:
-    """Directory holding one run's per-node result sidecars (P5)."""
-    return Path(output_dir) / "workflow_runs" / run_id
-
-
-def _project_inspectable(port_type: str, value: Any) -> Any:
-    """Strip a port value down to its inspectable (JSON-safe) projection.
-
-    The only heavy/non-serializable rider is an ``events`` value's
-    ``raw_results`` — per-frame scoring kept for the heatmap node, which can
-    carry numpy arrays and is large. Drop it; the rest of every inspectable type
-    is already JSON-safe (paths, counts, text, timestamps).
-    """
-    if port_type == "events" and isinstance(value, dict):
-        return {k: v for k, v in value.items() if k != "raw_results"}
-    return value
-
-
-def _filter_result_ports(
-    node_type_id: str, result: Any, allowed: frozenset[str]
-) -> dict[str, Any]:
-    """Keep only the result ports whose declared type is in ``allowed`` (projected)."""
-    if not isinstance(result, dict):
-        return {}
-    node_type = NODE_TYPES.get(node_type_id)
-    out_types = {p["name"]: p["type"] for p in (node_type or {}).get("outputs", [])}
-    kept: dict[str, Any] = {}
-    for port, val in result.items():
-        ptype = out_types.get(port)
-        if ptype in allowed:
-            kept[port] = _project_inspectable(ptype, val)
-    return kept
-
-
-def _inspectable_result(node_type_id: str, result: Any) -> dict[str, Any]:
-    """The UI-inspectable projection of a node result."""
-    return _filter_result_ports(node_type_id, result, _INSPECTABLE_PORT_TYPES)
-
-
-def inspectable_sidecar_view(payload: dict[str, Any]) -> dict[str, Any]:
-    """Project a stored sidecar payload down to its UI-inspectable ports.
-
-    Sidecars persist every JSON-safe port (for resume) plus a self-describing
-    ``__type__`` key; the run-history inspector should render only the
-    inspectable subset, exactly as it did before the widening.
-    """
-    node_type_id = str(payload.get("__type__", "") or "")
-    out_types = {
-        p["name"]: p["type"]
-        for p in (NODE_TYPES.get(node_type_id) or {}).get("outputs", [])
-    }
-    return {
-        port: val
-        for port, val in payload.items()
-        if port != "__type__" and out_types.get(port) in _INSPECTABLE_PORT_TYPES
-    }
-
-
-def write_node_sidecar(
-    output_dir: Path | str, run_id: str, node_id: str, node_type_id: str, result: Any
-) -> bool:
-    """Atomically write a node's JSON-safe result ports to its run sidecar.
-
-    Persists every ``_SIDECAR_PORT_TYPES`` port plus a self-describing
-    ``__type__`` key (consumed by resume + the read-time inspectable filter).
-    Returns True iff a sidecar file now exists for this node. JSON-sanitizes via
-    :func:`utils.sanitize_floats` (non-finite floats / numpy scalars). A bad
-    ``node_id`` (path separators) or an empty payload writes nothing and
-    returns False.
-    """
-    if not node_id or node_id != os.path.basename(node_id) or node_id in (".", ".."):
-        return False
-    payload = _filter_result_ports(node_type_id, result, _SIDECAR_PORT_TYPES)
-    if not payload:
-        return False
-    payload["__type__"] = node_type_id
-    path = run_results_dir(output_dir, run_id) / f"{node_id}.json"
-    tmp = path.with_suffix(".json.tmp")
-    try:
-        path.parent.mkdir(parents=True, exist_ok=True)
-        tmp.write_text(
-            json.dumps(utils.sanitize_floats(payload), ensure_ascii=False, indent=2),
-            encoding="utf-8",
-        )
-        os.replace(tmp, path)
-        return True
-    except (OSError, TypeError, ValueError) as exc:
-        utils.warning_print(f"workflow sidecar write failed ({node_id}): {exc}")
-        try:
-            tmp.unlink(missing_ok=True)
-        except OSError:
-            pass
-        return False
-
-
-# Collection nodes that pass an events value's ``raw_results`` through
-# unchanged (see ``_COLLECTION_KINDS["events"]["preserve"]`` / merge's concat).
-# The resume planner walks heatmap ancestry through these.
-_RAW_RESULTS_PRESERVING = frozenset(
-    {
-        "filter_events",
-        "partition_events",
-        "merge_events",
-        "limit_events",
-        "dedup_events",
-    }
-)
-
-
-def compute_resume_plan(
-    blueprint: dict[str, Any],
-    prior_node_states: dict[str, Any],
-    load_sidecar: Callable[[str], dict[str, Any] | None],
-) -> tuple[dict[str, dict[str, Any]], list[str]]:
-    """Plan a resume: which prior-run nodes can be reused as seeds vs. re-run.
-
-    Pure. ``load_sidecar(node_id)`` returns the stored sidecar payload (with
-    its ``__type__`` key) or None. Returns ``(seed_results, notes)`` where
-    ``seed_results`` feeds :class:`WorkflowRunner`'s ``seed_results`` and
-    ``notes`` carries human-readable degradation reasons.
-
-    A node re-runs when: it didn't complete in the prior run; its id/type
-    changed since (graph edited between runs); its sidecar is missing or
-    doesn't cover every declared output port (e.g. clipRecords producers,
-    which are never sidecar-persisted); OR any ancestor re-runs (fresh inputs
-    invalidate the cached output). Additionally, a re-running ``heatmap``
-    forces its completed events-ancestry to re-run too — sidecars project out
-    the heavy ``raw_results`` the heatmap consumes, so a seeded events
-    producer would make the resumed heatmap silently emit nothing.
-    """
-    nodes = [n for n in blueprint.get("nodes", []) if n.get("type") != NOTE_NODE_TYPE]
-    by_id = {n["id"]: n for n in nodes}
-    children: dict[str, list[str]] = {n["id"]: [] for n in nodes}
-    parents: dict[str, list[str]] = {n["id"]: [] for n in nodes}
-    for edge in blueprint.get("edges", []):
-        src, dst = edge.get("from"), edge.get("to")
-        if src in by_id and dst in by_id:
-            children[src].append(dst)
-            parents[dst].append(src)
-
-    notes: list[str] = []
-    rerun: set[str] = set()
-    seeds: dict[str, dict[str, Any]] = {}
-    for n in nodes:
-        nid = n["id"]
-        prior = prior_node_states.get(nid) or {}
-        if prior.get("status") != NODE_STATUS_COMPLETED:
-            rerun.add(nid)
-            continue
-        payload = load_sidecar(nid)
-        if not isinstance(payload, dict):
-            rerun.add(nid)
-            continue
-        if str(payload.get("__type__", "")) != str(n.get("type", "")):
-            rerun.add(nid)  # the node changed type since the prior run
-            continue
-        declared = (NODE_TYPES.get(str(n.get("type", ""))) or {}).get("outputs", [])
-        stored = {k: v for k, v in payload.items() if k != "__type__"}
-        if not declared or any(p["name"] not in stored for p in declared):
-            # Some output port wasn't persisted (non-JSON-safe type, or the
-            # executor omitted it) — downstream would see None; re-run instead.
-            rerun.add(nid)
-            continue
-        seeds[nid] = stored
-
-    def _close_under_descendants() -> None:
-        stack = list(rerun)
-        while stack:
-            for child in children.get(stack.pop(), []):
-                if child not in rerun:
-                    rerun.add(child)
-                    stack.append(child)
-
-    # Fixpoint: descendant closure and the heatmap raw_results rule feed each
-    # other (a forced ancestor invalidates its own seeded descendants).
-    while True:
-        before = len(rerun)
-        _close_under_descendants()
-        for n in nodes:
-            if n.get("type") != "heatmap" or n["id"] not in rerun:
-                continue
-            stack = list(parents.get(n["id"], []))
-            while stack:
-                pid = stack.pop()
-                if pid in rerun:
-                    continue
-                rerun.add(pid)
-                if by_id[pid].get("type") in _RAW_RESULTS_PRESERVING:
-                    stack.extend(parents.get(pid, []))
-        if len(rerun) == before:
-            break
-
-    seed_results = {nid: val for nid, val in seeds.items() if nid not in rerun}
-    if not seed_results:
-        notes.append("No prior results could be reused — running everything")
-    return seed_results, notes
-
-
-class WorkflowRunner:
-    """Executes one blueprint DAG; ``run()`` is the daemon-thread target.
-
-    Holds per-node state + an in-memory per-(node,port) result store. ``on_update``
-    is invoked on every status transition (immediately) and on throttled progress
-    ticks; the server wires it to the SSE notify. ``snapshot()`` returns the
-    JSON-safe run record (counts + status, large blobs stripped) for the API.
-    """
-
-    def __init__(
-        self,
-        run_id: str,
-        blueprint: dict[str, Any],
-        ctx: NodeContext,
-        on_update: Callable[[], None] | None = None,
-        participant: str = "",
-        batch_id: str = "",
-        triggered: bool = False,
-        trigger_type: str = "",
-        target_node_id: str = "",
-        seed_results: dict[str, dict[str, Any]] | None = None,
-        seed_note: str = "",
-    ) -> None:
-        self.run_id = run_id
-        self.blueprint_id = str(blueprint.get("id", "") or "")
-        # Partial run (P11): when set, only this node and its transitive ancestors
-        # execute; the rest are marked skipped. Empty → run the whole graph.
-        self.target_node_id = target_node_id
-        # Pre-seeded results: {node_id: result} for nodes whose output is
-        # already known — participant-independent sources a batch coordinator
-        # computed once (P3), or completed nodes reloaded from a prior run's
-        # sidecars on resume. A seeded node is stored as if it ran, skipping
-        # its executor. ``seed_note`` (resume) is surfaced on each seeded node.
-        self._seed_results = seed_results or {}
-        self._seed_note = seed_note
-        # Batch identity (P3): empty for a normal single run; a child run carries
-        # its participant + parent batch id so the snapshot can be grouped.
-        self.participant = participant
-        self.batch_id = batch_id
-        # Auto-run triggers (P6 + chaining): True when this run was launched by
-        # the watcher (surfaced as a badge in the run history); ``trigger_type``
-        # records which trigger fired it (new_video / transcript_complete / …).
-        self.triggered = triggered
-        self.trigger_type = trigger_type
-        # Sticky notes are canvas annotations, not executable nodes — drop them
-        # before node_states is built so they never run, fail as "No executor",
-        # or pad the snapshot's node counts.
-        self.nodes = [
-            n for n in blueprint.get("nodes", []) if n.get("type") != NOTE_NODE_TYPE
-        ]
-        self.edges = list(blueprint.get("edges", []))
-        self.ctx = ctx
-        self.on_update = on_update or (lambda: None)
-        self._nodes_by_id = {n["id"]: n for n in self.nodes}
-        self.node_states: dict[str, dict[str, Any]] = {
-            n["id"]: {
-                "status": NODE_STATUS_QUEUED,
-                "progress": 0.0,
-                "error": None,
-                # Non-fatal note for a degraded-but-completed node (Ollama down,
-                # nothing wired, an adapter that couldn't coerce) — distinct from
-                # ``error`` (which means FAILED). Surfaced in the run history.
-                "note": None,
-                "started_at": None,
-                "completed_at": None,
-            }
-            for n in self.nodes
-        }
-        self._results: dict[str, dict[str, Any]] = {}
-        # Node ids with an inspectable result sidecar on disk (P5); surfaced as
-        # ``hasResult`` in the snapshot so the UI knows it can fetch on demand.
-        self._sidecars: set[str] = set()
-        self.status = RUN_STATUS_QUEUED
-        self.started_at: str | None = None
-        self.completed_at: str | None = None
-        self._lock = threading.Lock()
-        self._last_notify = 0.0
-
-    # ---- control ----
-
-    def cancel(self) -> None:
-        """Signal the run-wide cancel event (checked between nodes + by executors)."""
-        self.ctx.cancel_event.set()
-
-    # ---- graph helpers ----
-
-    def _incoming(self, node_id: str) -> list[dict[str, Any]]:
-        return [e for e in self.edges if e.get("to") == node_id]
-
-    def _deps(self, node_id: str) -> set[str]:
-        deps: set[str] = set()
-        for edge in self._incoming(node_id):
-            src = edge.get("from")
-            if isinstance(src, str) and src in self._nodes_by_id:
-                deps.add(src)
-        return deps
-
-    def _ancestors_inclusive(self, node_id: str) -> set[str]:
-        """``node_id`` plus everything it transitively depends on (for partial runs)."""
-        seen: set[str] = set()
-        stack = [node_id]
-        while stack:
-            nid = stack.pop()
-            if nid in seen:
-                continue
-            seen.add(nid)
-            stack.extend(self._deps(nid))
-        return seen
-
-    def _gate_blocks(self, node_id: str) -> bool:
-        """True if ``node_id`` is a gate that completed with ``pass`` False."""
-        node = self._nodes_by_id.get(node_id)
-        if not node or node.get("type") not in ("gate", "gate_collection"):
-            return False
-        return (self._results.get(node_id) or {}).get("pass") is False
-
-    def _should_skip(self, node_id: str) -> bool:
-        """Decide whether ``node_id`` is skipped before it runs.
-
-        Skipped when a *required* input's producer failed/skipped (the node can't
-        run without it) or when a gate gating this node blocks it. A dead producer
-        feeding an *optional* input is tolerated — the node runs with that input
-        absent (executors already read ``inputs.get(...)``), so a muted or broken
-        branch of a ``merge`` (or any optional input) doesn't sink the whole
-        downstream.
-        """
-        node = self._nodes_by_id.get(node_id)
-        node_type = node.get("type") if node else None
-        for edge in self._incoming(node_id):
-            dep = edge.get("from")
-            if not isinstance(dep, str) or dep not in self._nodes_by_id:
-                continue
-            status = self.node_states[dep]["status"]
-            out_type = _port_type(
-                self._nodes_by_id[dep].get("type"), edge.get("fromPort"), "out"
-            )
-            if out_type == "control":
-                # A gate edge: skip if the gate can't pass us through — it blocked
-                # (``pass`` False) or it never completed (failed/skipped).
-                if status in (NODE_STATUS_FAILED, NODE_STATUS_SKIPPED):
-                    return True
-                if status == NODE_STATUS_COMPLETED and self._gate_blocks(dep):
-                    return True
-                continue
-            # A data edge: only a *required* input's dead producer forces a skip.
-            if status in (
-                NODE_STATUS_FAILED,
-                NODE_STATUS_SKIPPED,
-            ) and not _port_optional(node_type, edge.get("toPort")):
-                return True
-        return False
-
-    def _gather_inputs(self, node: dict[str, Any]) -> tuple[dict[str, Any], list[str]]:
-        """Map upstream results onto this node's input ports, applying adapters.
-
-        Returns ``(inputs, notes)``. ``notes`` carries any adapter-failure
-        messages so the runner can surface them on the node — a coercion that
-        raises otherwise degrades the input to ``None`` invisibly (only a server
-        log). Control edges (a gate's ``control`` output) establish a dependency
-        but carry no data, so they are excluded here — they never clobber a real
-        input.
-        """
-        inputs: dict[str, Any] = {}
-        notes: list[str] = []
-        for edge in self._incoming(node["id"]):
-            src_id = edge.get("from")
-            to_port = edge.get("toPort")
-            from_port = edge.get("fromPort")
-            if not (
-                isinstance(src_id, str)
-                and isinstance(to_port, str)
-                and isinstance(from_port, str)
-            ):
-                continue
-            src_node = self._nodes_by_id.get(src_id)
-            if src_node is None:
-                continue
-            out_type = _port_type(src_node["type"], from_port, "out")
-            if out_type == "control":
-                continue
-            in_type = _port_type(node["type"], to_port, "in")
-            value = (self._results.get(src_id) or {}).get(from_port)
-            if out_type is not None and in_type is not None and out_type != in_type:
-                adapter = ADAPTERS.get((out_type, in_type))
-                if adapter is not None:
-                    try:
-                        value = adapter(value)
-                    except Exception as exc:  # adapter failure → empty input + note
-                        utils.warning_print(
-                            f"workflow adapter {out_type}->{in_type} failed: {exc}"
-                        )
-                        notes.append(f"Couldn't convert {out_type} to {in_type}")
-                        value = None
-            inputs[to_port] = value
-        return inputs, notes
-
-    # ---- state + notify ----
-
-    def _set_node(self, node_id: str, **changes: Any) -> None:
-        with self._lock:
-            self.node_states[node_id].update(changes)
-
-    def _notify(self, force: bool = False) -> None:
-        now = time.monotonic()
-        if force or now - self._last_notify >= _PROGRESS_NOTIFY_INTERVAL:
-            self._last_notify = now
-            try:
-                self.on_update()
-            except Exception:
-                pass
-
-    def _make_progress(self, node_id: str) -> Callable[[float], None]:
-        def _on_progress(fraction: float) -> None:
-            with self._lock:
-                self.node_states[node_id]["progress"] = max(
-                    0.0, min(1.0, float(fraction))
-                )
-            self._notify()
-
-        return _on_progress
-
-    # ---- run ----
-
-    def run(self) -> None:
-        """Execute the DAG in topological order. Safe to call once, on a thread."""
-        self.status = RUN_STATUS_RUNNING
-        self.started_at = _now_iso()
-        self._notify(force=True)
-        try:
-            order = topo_order(self.nodes, self.edges)
-        except WorkflowCycleError:
-            self.status = RUN_STATUS_FAILED
-            self.completed_at = _now_iso()
-            self._notify(force=True)
-            return
-
-        # Partial run: keep only the target node and its ancestors; the rest are
-        # skipped up front (they never execute and don't block completion).
-        if self.target_node_id and self.target_node_id in self._nodes_by_id:
-            keep = self._ancestors_inclusive(self.target_node_id)
-            for nid in order:
-                if nid not in keep:
-                    self._set_node(
-                        nid, status=NODE_STATUS_SKIPPED, completed_at=_now_iso()
-                    )
-            order = [nid for nid in order if nid in keep]
-
-        for node_id in order:
-            node = self._nodes_by_id[node_id]
-            if self.ctx.cancel_event.is_set():
-                self._set_node(
-                    node_id, status=NODE_STATUS_SKIPPED, completed_at=_now_iso()
-                )
-                continue
-            # Seeded result (batch precompute or resume): the node's output is
-            # authoritatively known — store it as if it just ran, skipping its
-            # executor. Checked BEFORE the mute/skip gates: a resume seed for a
-            # completed node must survive even when a (re-run) parent upstream
-            # is currently marked skipped/muted — the prior run already proved
-            # this node's output. (Batch seeds are parentless sources, so the
-            # ordering change is behavior-neutral for them.)
-            if node_id in self._seed_results:
-                seeded = self._seed_results[node_id]
-                with self._lock:
-                    self._results[node_id] = seeded
-                if write_node_sidecar(
-                    self.ctx.output_dir, self.run_id, node_id, node["type"], seeded
-                ) and _inspectable_result(node["type"], seeded):
-                    with self._lock:
-                        self._sidecars.add(node_id)
-                self._set_node(
-                    node_id,
-                    status=NODE_STATUS_COMPLETED,
-                    progress=1.0,
-                    completed_at=_now_iso(),
-                    note=self._seed_note or None,
-                )
-                self._notify(force=True)
-                continue
-
-            # A muted node is skipped intrinsically; _should_skip then propagates
-            # SKIPPED to its whole downstream subtree (same as a blocking gate).
-            if node.get("disabled"):
-                self._set_node(
-                    node_id, status=NODE_STATUS_SKIPPED, completed_at=_now_iso()
-                )
-                self._notify(force=True)
-                continue
-            if self._should_skip(node_id):
-                self._set_node(
-                    node_id, status=NODE_STATUS_SKIPPED, completed_at=_now_iso()
-                )
-                self._notify(force=True)
-                continue
-
-            inputs, input_notes = self._gather_inputs(node)
-            params = node.get("params", {}) or {}
-            executor = NODE_TYPES.get(node["type"], {}).get("execute")
-            self._set_node(
-                node_id, status=NODE_STATUS_RUNNING, started_at=_now_iso(), progress=0.0
-            )
-            self._notify(force=True)
-            if executor is None:
-                self._set_node(
-                    node_id,
-                    status=NODE_STATUS_FAILED,
-                    error=f"No executor for node type: {node.get('type')}",
-                    completed_at=_now_iso(),
-                )
-                self._notify(force=True)
-                continue
-
-            self.ctx.on_progress = self._make_progress(node_id)
-            try:
-                result = executor(self.ctx, inputs, params)
-                result = result if isinstance(result, dict) else {}
-                # A reserved ``__note__`` key lets an executor flag a non-fatal
-                # degraded outcome (e.g. Ollama unavailable, nothing wired) that
-                # still completes — surfaced on the node, never stored as a result
-                # port. Merge it with any adapter-coercion notes from gathering.
-                notes = list(input_notes)
-                exec_note = result.pop("__note__", None)
-                if exec_note:
-                    notes.append(str(exec_note))
-                with self._lock:
-                    self._results[node_id] = result
-                # Persist the JSON-safe result ports (resume reloads them; the
-                # run-history UI fetches the inspectable subset on demand even
-                # after this runner is evicted from memory). ``hasResult`` only
-                # advertises sidecars with something the inspector can render.
-                if write_node_sidecar(
-                    self.ctx.output_dir, self.run_id, node_id, node["type"], result
-                ) and _inspectable_result(node["type"], result):
-                    with self._lock:
-                        self._sidecars.add(node_id)
-                self._set_node(
-                    node_id,
-                    status=NODE_STATUS_COMPLETED,
-                    progress=1.0,
-                    completed_at=_now_iso(),
-                    note="; ".join(notes) if notes else None,
-                )
-            except Exception as exc:
-                self._set_node(
-                    node_id,
-                    status=NODE_STATUS_FAILED,
-                    error=str(exc),
-                    completed_at=_now_iso(),
-                )
-            self._notify(force=True)
-
-        if self.ctx.cancel_event.is_set():
-            self.status = RUN_STATUS_CANCELLED
-        elif any(s["status"] == NODE_STATUS_FAILED for s in self.node_states.values()):
-            self.status = RUN_STATUS_FAILED
-        else:
-            self.status = RUN_STATUS_COMPLETED
-        self.completed_at = _now_iso()
-        self._notify(force=True)
-
-    # ---- serialization ----
-
-    def snapshot(self) -> dict[str, Any]:
-        """JSON-safe run record for the API/SSE/manifest (counts + status only)."""
-        with self._lock:
-            node_states = {
-                nid: {
-                    **{k: v for k, v in st.items() if not k.startswith("_")},
-                    "hasResult": nid in self._sidecars,
-                }
-                for nid, st in self.node_states.items()
-            }
-            results = {
-                nid: _node_result_summary(res) for nid, res in self._results.items()
-            }
-        return {
-            "id": self.run_id,
-            "blueprintId": self.blueprint_id,
-            "batchId": self.batch_id,
-            "participant": self.participant,
-            "triggered": self.triggered,
-            "triggerType": self.trigger_type,
-            "status": self.status,
-            "nodeStates": node_states,
-            "results": results,
-            "startedAt": self.started_at,
-            "completedAt": self.completed_at,
-        }
